'Refactor sony store methods'
